WIP: new API surface - types
