+++ src/index.ts
@@ -1,5 +1,3 @@
-import { Texture } from './texture/Texture';
 import { TonyGL } from './TonyGL';
 
 export { TonyGL };
-export type { Texture };
