+++ src/materials/BaseMaterialFactory.ts
@@ -1,39 +1,44 @@
-import { Renderer } from '../renderer/configureRenderer';
-import { CreateUniformBufferFunction, UniformBuffer, UniformObject } from '../core/UniformBufferFactory';
-import type { MaterialType, BaseMaterial } from './MaterialsFactory';
+import { CreateUniformBufferFunction, UniformBuffer, UniformObject } from '../core/core.types';
+import type { Renderer } from '../renderer/renderer.types';
+import type { MaterialType, BaseMaterial } from './materials.types';
+
+export const enum MaterialFlags {
+  None = 0,
+  Alpha = 1 << 0,
+  Normal = 1 << 1,
+  Albedo = 1 << 2,
+}
+
+export type BaseMaterialOptions = {
+  type: MaterialType;
+  shader: string;
+  uniforms?: UniformObject;
+  transparent?: boolean;
+  doubleSided?: boolean;
+  depthWrite?: boolean;
+  buildBindGroupEntries: (materialUniformsBuffer: UniformBuffer | null) => GPUBindGroupEntry[];
+};
 
 export function BaseMaterialFactory(renderer: Renderer, createUniformBuffer: CreateUniformBufferFunction) {
-  function createBaseMaterial(options: {
-    type: MaterialType;
-    shader: string;
-    uniforms?: UniformObject;
-    transparent?: boolean;
-    doubleSided?: boolean;
-    depthWrite?: boolean;
-    buildEntries: (materialUniformsBuffer: UniformBuffer | null) => GPUBindGroupEntry[];
-  }): BaseMaterial {
+  function createBaseMaterial(options: BaseMaterialOptions): BaseMaterial {
     const id = crypto.randomUUID();
     const type = options.type;
-    let shader = options.shader;
+    const materialUniformsBuffer = options.uniforms ? createUniformBuffer(options.uniforms) : null;
+    const shaderIdentifier =
+      type === 'custom' ? renderer.shaderLibrary.buildCustomShader({ shader: options.shader, id }) : options.shader;
 
-    if (type === 'custom') {
-      shader = renderer.shaderLibrary.buildCustomShader({ shader: options.shader, id });
-    }
-
-    const shaderRef = renderer.shaderLibrary.getShader(shader);
-
-    if (!shaderRef) {
+    const cachedShader = renderer.shaderLibrary.getShader(shaderIdentifier);
+    if (!cachedShader) {
       throw new Error('Could not find shader code for material.');
     }
 
-    const shaderModule = shaderRef.shaderModule;
-    const materialUniformsBuffer = options.uniforms ? createUniformBuffer(options.uniforms) : null;
     const materialBindGroupLayout = renderer.bindGroupLayouts.materialBindGroupLayouts.get(type);
     if (!materialBindGroupLayout) {
       throw new Error(`Material bind group layout missing for type: ${type}`);
     }
 
-    const entries = options.buildEntries(materialUniformsBuffer);
+    const entries = options.buildBindGroupEntries(materialUniformsBuffer);
+
     const materialUniformsBindGroup =
       entries.length > 0
         ? renderer.device.createBindGroup({
@@ -45,8 +50,8 @@ export function BaseMaterialFactory(renderer: Renderer, createUniformBuffer: Cre
     const material: BaseMaterial = {
       id,
       type,
-      shader,
-      shaderModule,
+      shader: shaderIdentifier,
+      shaderModule: cachedShader.shaderModule,
       materialUniformsBuffer,
       materialUniformsBindGroup,
       transparent: options.transparent ?? false,
